added branchResultset() to dynamic view
While you can always access the dynamic view's internal 'resultset', you
could in theory mess up the state of the view by applying filters to it
instead of the view.  Therefore this is just a simple helper method to
give you a clone of the resultset for branch query possibilities.