Added service to loki-angular.js called Lokiwork
The Lokiwork service helps an angular programmer quickly get up and running with Lokijs.  This service automatically builds your database if it's the first run or it detects the databases were wiped (which can happen as you know on a mobile device).  The programmer doesn't have to worry about this.  It also allows the programmer to use json files in an easy way without having to use special plugins to access local file storage or having to embed json code directly within your services.  It also helps facilitate switching between databases automatically for the developer when needed, loading collections into databases, etc.  It's ideal for someone that wants to quickly get up and running using loki having to worry about why the db context is now null all of a sudden or why the callback isn't getting called correctly or how to build the database template in an elegant way without requiring the developer spend ours searching the web for a better approach.  I have a help doc describing how to use if anyone ever needs it.