Call loadDatabase callback with parsing error
When the database's JSON cannot be parsed back, we have to call the
callback function with the error.  This is important as the
loadDatabaseCallback can be called asynchronously, and in that case, we
have no means to intercept the error if cFun is not called.

With this fix, the error can be handled in the callback function that is
passed to Loki.loadDatabase().