don't throw when emitting events without listeners
This brings the EventEmitter implementation in line with the NodeJS API.

This always caused errors (e. g. the emit("load") in Loki.prototype.loadDatabase) but since the switch to Promises, those are not swallowed anymore but instead reported as unhandled exceptions.

Since this behaviour clearly seems to be by design it might be, that I missed something here.