safeguard for binary indices with date values
since dates serialize as strings which have a different sort order, this
could have corrupted a binary index.  To ensure this does not happen, if
you insert or update an object with binary indices, we will convert any
date values in those indexed properties to epoch time.  If you do not
wish to serialize you can set collection option
'serializableIndices':false in constructor and we will leave as dates.