#594 fix : do not throw adapter errors when db doesn't exist
Partitioning adapter as well as LokiMemoryAdapter were throwing an error when attempting to load a database that doesn't (yet) exist.  Now if adapter's loadDatabase method returns falsy (like null) value it means db doesnt (yet) exist.  If it returns instance of error, we raise error.  If it returns string we parse and load from object.  If it returns object we load from object.