fix for #639 : binary index in batch updates when not cloning
Batch updates will now temporarily disable adaptive indices (when not cloning) for the duration of the batch update if binary indices are defined.  We will do a full lazy rebuild when batch update completes and re-enable adaptive indices.  This fixes edge case where binary indices could become corrupt.  If your batches are small or medium sized it will likely be faster to update docs individually, as they are modified instead.